requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.109.0
fastapi-sso==0.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
mysql-connector-python==8.3.0
oauthlib==3.2.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
